Use src layout and modernize CI/CD (#43)
The src layout is recommended to make sure tests and packaging does not
accidentally include extraneous files.

We also update all of CI/CD to run on uv for speed and easier
cross-platform support.

* Rename CI/CD file and use compact matrix names

This makes it easier to see at a glance the platform and python version.

* Add uv.lock for reproducible builds

It is considered good practice to version control the lock file.
The CI/CD script will raise an error if the pyproject.toml file is not
compatible.

* Remove custom shell and run CI on draft PR

This allows code checks and tests to run successfully on Windows.

Running CI on draft PRs is also recommended so no need to disable them,
and we also resize CI section headers to improve diffs.

* Test all supported Python versions on Linux
* Install both uv and Python using the same action
* Remove unused and outdated files

The guidelines listed in contributing.md have become outdated and
reference repositories which are no longer actively maintained.

Removing for now, but we may consider adding a contribution section to
the README file later.

* Update pre-commit hooks
* Apply formatting fixes
* Exclude repo config and data files from pre-commit
* Run code coverage only for latest Python version
* Avoid running tests twice on coverage report runs