Add missing type hints for api module
Runtime type checks were removed for validating invalid string, since in
principle any object could be passed at runtime, not just an invalid
string.

If type checking is relied on anyway to determine a value is a str, then
the same checker will validate it belongs to the valid literal set.